+++ source/extensions/omni.isaac.orbit/docs/CHANGELOG.rst
@@ -1,6 +1,19 @@
 Changelog
 ---------
 
+0.10.11 (2024-01-08)
+~~~~~~~~~~~~~~~~~~~~
+
+Fixed
+^^^^^
+
+* Fixed attribute error raised when calling the :class:`omni.isaac.orbit.envs.mdp.TerrainBasedPositionCommand`
+  command term.
+* Added a dummy function in :class:`omni.isaac.orbit.terrain.TerrainImporter` that returns environment
+  origins as terrain-aware sampled targets. This function should be implemented by child classes based on
+  the terrain type.
+
+
 0.10.10 (2023-12-21)
 ~~~~~~~~~~~~~~~~~~~~
 
@@ -10,6 +23,7 @@ Fixed
 * Fixed reliance on non-existent ``Viewport`` in :class:`omni.isaac.orbit.sim.SimulationContext` when loading livestreaming
   by ensuring that the extension ``omni.kit.viewport.window`` is enabled in :class:`omni.isaac.orbit.app.AppLauncher` when
   livestreaming is enabled
+
 
 0.10.9 (2023-12-21)
 ~~~~~~~~~~~~~~~~~~~